+++ src/main/org/openjfx/model/SerialPortService.java
@@ -1,14 +1,14 @@
 package org.openjfx.model;
 
 import com.fazecast.jSerialComm.SerialPort;
-import org.openjfx.model.settings.MasterSettings;
 import org.openjfx.model.settings.PortSettings;
-import org.openjfx.model.settings.SlaveSettings;
 
 public class SerialPortService {
 
-
-    public SerialPort GetInitializedPort(PortSettings settings){
+    /*TODO
+     * jeszcze ta kontrola przeplywu nie jest nigdzie uwzgledniona
+     */
+    public SerialPort getInitializedPort(PortSettings settings){
         SerialPort port = SerialPort.getCommPort(settings.getPortDescription());
         if(port!=null) {
             System.out.println("otrzymano port:" + port.getDescriptivePortName());
@@ -28,25 +28,26 @@ public class SerialPortService {
             System.out.println("Nie udało sie dostać żadnego portu => JESTESMY W DUPIE");
         }
         return port;
-
-
-
     }
 
-    public void SendString(SerialPort port,String message){
+    /*TODO
+     * dodac dodanie odpowiedniego terminatora na koncu
+     */
+    public void sendString(SerialPort port,String message){
         port.openPort();
-
         int bytesWritten = port.writeBytes(message.getBytes(), message.length());
         if(bytesWritten >0)  System.out.println("Coś sie wysłało, jest dobrze (y)");
-
+        else System.out.println("Nic sie nie wysłało");
         port.closePort();
-
-
     }
 
-    public void ReciveData(SerialPort port){
-        port.openPort();
 
+    /*TODO
+     * nie trzeba sprawdzac tez terminatora?
+     */
+    public String reciveData(SerialPort port){
+        port.openPort();
+        String data;
         try {
             while (true)
             {
@@ -54,15 +55,25 @@ public class SerialPortService {
                     Thread.sleep(20);
 
                 byte[] readBuffer = new byte[port.bytesAvailable()];
-                String data = readBuffer.toString();
+                data = readBuffer.toString();
                 int numRead = port.readBytes(readBuffer, readBuffer.length);
                 System.out.println("Przeczytano " + numRead + " bajtow.");
                 System.out.println("Wiadomośc: " + data);
             }
-        } catch (Exception e) { e.printStackTrace(); }
-        port.closePort();
+        }
+        catch (Exception e) {
+            //e.printStackTrace();
+        data = "could not read";
+        }
+        finally {
+            port.closePort();
+        }
+        return data;
     }
 
+    /*TODO
+     * dodac ping
+     */
     public boolean ping(){
         return true;
     }
