+++ src/main/org/openjfx/view/receive_mode.fxml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.text.*?>
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.openjfx.controller.ReceiveController">
+
+    <children>
+        <TextArea fx:id="receivedText" layoutX="35.0" layoutY="40.0" prefHeight="325.0" prefWidth="362.0" />
+        <Button fx:id="btn_send" layoutX="469.0" layoutY="46.0" mnemonicParsing="false" onAction="#receive" text="Receive" />
+      <Button fx:id="btn_ping" layoutX="479.0" layoutY="104.0" mnemonicParsing="false" onAction="#ping" text="PING" />
+      <Button fx:id="btn_return" layoutX="438.0" layoutY="161.0" mnemonicParsing="false" onAction="#returnToSettings" text="return to settings" />
+      <Label layoutX="44.0" layoutY="14.0" prefHeight="15.0" prefWidth="206.0" text="Received text:" />
+    </children>
+
+</AnchorPane>
