+++ src/main/org/openjfx/view/send_mode.fxml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.text.*?>
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.openjfx.controller.SendController">
+   <children>
+      <TextArea fx:id="inputText" layoutX="35.0" layoutY="51.0" prefHeight="314.0" prefWidth="362.0" />
+      <Button fx:id="btn_send" layoutX="473.0" layoutY="72.0" mnemonicParsing="false" onAction="#send" text="Send" />
+      <Button fx:id="btn_ping" layoutX="474.0" layoutY="133.0" mnemonicParsing="false" onAction="#ping" text="PING" />
+      <Button fx:id="btn_return" layoutX="433.0" layoutY="196.0" mnemonicParsing="false" onAction="#returnToSettings" text="return to settings" />
+      <Label layoutX="35.0" layoutY="28.0" prefHeight="15.0" prefWidth="223.0" text="Text to send:" />
+   </children>
+</AnchorPane>
